+++ src/fix_boundary_sources.cpp
@@ -18,12 +18,6 @@
 #include "config.h"
 
 #ifdef HAVE_MPI
-#ifdef NEED_UNDEF_SEEK_FOR_MPI
-// undef'ing SEEK_* is needed for MPICH, possibly other MPI versions
-#undef SEEK_SET
-#undef SEEK_END
-#undef SEEK_CUR
-#endif
 #include <mpi.h>
 #endif