+++ src/step_db.cpp
@@ -103,11 +103,11 @@ bool fields_chunk::step_db(field_type ft) {
                  and sums, but you get the idea). */
               if (!f_rderiv_int) f_rderiv_int = new realnum[gv.ntot()];
               realnum ir0 = gv.origin_r() * gv.a + 0.5 * gv.iyee_shift(c_p).in_direction(R);
-              for (int iz = 0; iz <= gv.nz(); ++iz)
-                f_rderiv_int[iz] = 0;
+              memset(f_rderiv_int, 0, sizeof(realnum) * (gv.nz() + 1));
               int sr = gv.nz() + 1;
               for (int ir = 1; ir <= gv.nr(); ++ir) {
                 realnum rinv = 1.0 / ((ir + ir0) - 0.5);
+                IVDEP
                 for (int iz = 0; iz <= gv.nz(); ++iz) {
                   ptrdiff_t idx = ir * sr + iz;
                   f_rderiv_int[idx] =
@@ -200,16 +200,15 @@ bool fields_chunk::step_db(field_type ft) {
             2 * m * (1 - 2 * cmp) * (1 - 2 * (ft == B_stuff)) * (1 - 2 * (d_c == R)) * Courant;
 
         // 8 special cases of the same loop (sigh):
-        if (siginv) {    // PML in f update
+        if (siginv) { // PML in f update
+          KSTRIDE_DEF(dsig, k, is, gv);
           if (siginvu) { // PML + fu
-            if (cndinv)  // PML + fu + conductivity
+            KSTRIDE_DEF(dsigu, ku, is, gv);
+            if (cndinv) // PML + fu + conductivity
               //////////////////// MOST GENERAL CASE //////////////////////
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsig, k, is, gv);
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_k;
-                KSTRIDE_DEF(dsigu, ku, is, gv);
                 DEF_ku;
                 realnum df, dfcnd = rinv * g[i] * cndinv[i];
                 fcnd[i] += dfcnd;
@@ -218,12 +217,9 @@ bool fields_chunk::step_db(field_type ft) {
               }
             /////////////////////////////////////////////////////////////
             else // PML + fu - conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsig, k, is, gv);
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_k;
-                KSTRIDE_DEF(dsigu, ku, is, gv);
                 DEF_ku;
                 realnum df, dfcnd = rinv * g[i];
                 fu[i] += (df = dfcnd * siginv[k]);
@@ -232,20 +228,16 @@ bool fields_chunk::step_db(field_type ft) {
           }
           else {        // PML - fu
             if (cndinv) // PML - fu + conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsig, k, is, gv);
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_k;
                 realnum dfcnd = rinv * g[i] * cndinv[i];
                 fcnd[i] += dfcnd;
                 the_f[i] += dfcnd * siginv[k];
               }
             else // PML - fu - conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsig, k, is, gv);
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_k;
                 realnum dfcnd = rinv * g[i];
                 the_f[i] += dfcnd * siginv[k];
@@ -254,21 +246,18 @@ bool fields_chunk::step_db(field_type ft) {
         }
         else {           // no PML in f update
           if (siginvu) { // no PML + fu
-            if (cndinv)  // no PML + fu + conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsigu, ku, is, gv);
+            KSTRIDE_DEF(dsigu, ku, is, gv);
+            if (cndinv) // no PML + fu + conductivity
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_ku;
                 realnum df = rinv * g[i] * cndinv[i];
                 fu[i] += df;
                 the_f[i] += siginvu[ku] * df;
               }
             else // no PML + fu - conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
-                KSTRIDE_DEF(dsigu, ku, is, gv);
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 DEF_ku;
                 realnum df = rinv * g[i];
                 fu[i] += df;
@@ -277,15 +266,13 @@ bool fields_chunk::step_db(field_type ft) {
           }
           else {        // no PML - fu
             if (cndinv) // no PML - fu + conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 the_f[i] += rinv * g[i] * cndinv[i];
               }
             else // no PML - fu - conductivity
-              LOOP_OVER_VOL_OWNED0(gv, cc, i) {
-                IVEC_LOOP_ILOC(gv, here);
-                realnum rinv = the_m / here.r();
+              PLOOP_OVER_VOL_OWNED0(gv, cc, i) {
+                realnum rinv = the_m / (loop_is2 + 2 * loop_i2);
                 the_f[i] += rinv * g[i];
               }
           }
@@ -417,10 +404,13 @@ bool fields_chunk::step_db(field_type ft) {
           if (ft == D_stuff)
             for (int r = 0; r <= gv.nr() && r < rmax; r++) {
               const int ir = r * (nz + 1);
+              if (f[Dr][cmp]) ZERO_Z(f[Dr][cmp] + ir);
               ZERO_Z(f[Dp][cmp] + ir);
               ZERO_Z(f[Dz][cmp] + ir);
+              if (f_cond[Dr][cmp]) ZERO_Z(f_cond[Dr][cmp] + ir);
               if (f_cond[Dp][cmp]) ZERO_Z(f_cond[Dp][cmp] + ir);
               if (f_cond[Dz][cmp]) ZERO_Z(f_cond[Dz][cmp] + ir);
+              if (f_u[Dr][cmp]) ZERO_Z(f_u[Dr][cmp] + ir);
               if (f_u[Dp][cmp]) ZERO_Z(f_u[Dp][cmp] + ir);
               if (f_u[Dz][cmp]) ZERO_Z(f_u[Dz][cmp] + ir);
             }
@@ -428,8 +418,14 @@ bool fields_chunk::step_db(field_type ft) {
             for (int r = 0; r <= gv.nr() && r < rmax; r++) {
               const int ir = r * (nz + 1);
               ZERO_Z(f[Br][cmp] + ir);
+              if (f[Bp][cmp]) ZERO_Z(f[Bp][cmp] + ir);
+              if (f[Bz][cmp]) ZERO_Z(f[Bz][cmp] + ir);
               if (f_cond[Br][cmp]) ZERO_Z(f_cond[Br][cmp] + ir);
+              if (f_cond[Bp][cmp]) ZERO_Z(f_cond[Bp][cmp] + ir);
+              if (f_cond[Bz][cmp]) ZERO_Z(f_cond[Bz][cmp] + ir);
               if (f_u[Br][cmp]) ZERO_Z(f_u[Br][cmp] + ir);
+              if (f_u[Bp][cmp]) ZERO_Z(f_u[Bp][cmp] + ir);
+              if (f_u[Bz][cmp]) ZERO_Z(f_u[Bz][cmp] + ir);
             }
         }
         else {
@@ -440,17 +436,26 @@ bool fields_chunk::step_db(field_type ft) {
              1/(|m|+0.5) is purely empirical (no theory yet), and I'm not
              sure how universal it is.  Makes higher m's more expensive. */
           if (ft == D_stuff) {
+            if (f[Dr][cmp]) ZERO_Z(f[Dr][cmp]);
             ZERO_Z(f[Dp][cmp]);
             ZERO_Z(f[Dz][cmp]);
+            if (f_cond[Dr][cmp]) ZERO_Z(f_cond[Dr][cmp]);
             if (f_cond[Dp][cmp]) ZERO_Z(f_cond[Dp][cmp]);
             if (f_cond[Dz][cmp]) ZERO_Z(f_cond[Dz][cmp]);
+            if (f_u[Dr][cmp]) ZERO_Z(f_u[Dr][cmp]);
             if (f_u[Dp][cmp]) ZERO_Z(f_u[Dp][cmp]);
             if (f_u[Dz][cmp]) ZERO_Z(f_u[Dz][cmp]);
           }
           else {
             ZERO_Z(f[Br][cmp]);
+            if (f[Bp][cmp]) ZERO_Z(f[Bp][cmp]);
+            if (f[Bz][cmp]) ZERO_Z(f[Bz][cmp]);
             if (f_cond[Br][cmp]) ZERO_Z(f_cond[Br][cmp]);
+            if (f_cond[Bp][cmp]) ZERO_Z(f_cond[Bp][cmp]);
+            if (f_cond[Bz][cmp]) ZERO_Z(f_cond[Bz][cmp]);
             if (f_u[Br][cmp]) ZERO_Z(f_u[Br][cmp]);
+            if (f_u[Bp][cmp]) ZERO_Z(f_u[Bp][cmp]);
+            if (f_u[Bz][cmp]) ZERO_Z(f_u[Bz][cmp]);
           }
         }
       }